refactor(enum): extract ignore patterns into reusable sub-package
Move CompileIgnorePatterns, ignore.conf, and tests into
pkg/enum/ignore so external callers can import the package
directly. The old enum.CompileIgnorePatterns is kept as a
deprecated thin wrapper for backwards compatibility.
